requirements.txt
pandas
numpy
ta
tensorflow
scikit-learn
joblib
python-binance
python-dotenv
requests